+++ CHANGELOG.md
@@ -80,11 +80,10 @@ Better execution report in markdown format
 Better execution reports
 
 -   Filter out voids in `executionReportJsonToString`
--   Add timing information to `ExecutionReportJson` <-[🎡]
--   Add money cost information to `ExecutionReportJson` <-[🎡]
+-   Add timing information to `ExecutionReportJson`
+-   Add money cost information to `ExecutionReportJson`
 -   Escape code blocks in markdown
 -   Do not export `replaceParameters` utility function
-
 
 ## 0.32.0 _(2024-02-XX)_